kafka_consumer/requirements.txt
kafka-python
pandas